move the types used in C programs to c_types.h
The rationale is that types of a particular group (say bitvectors or
pointers) are often used together with the expressions that are specific for
them.

The goal is that std_types.h will only contain the basic types, such as
bool_typet.
