+++ src/goto-programs/goto_trace.cpp
@@ -16,8 +16,8 @@ Author: Daniel Kroening
 #include <ostream>
 
 #include <util/arith_tools.h>
-#include <util/bitvector_types.h>
 #include <util/byte_operators.h>
+#include <util/c_types.h>
 #include <util/format_expr.h>
 #include <util/merge_irep.h>
 #include <util/range.h>
