+++ src/solvers/flattening/c_bit_field_replacement_type.h
@@ -10,7 +10,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #ifndef CPROVER_SOLVERS_FLATTENING_C_BIT_FIELD_REPLACEMENT_TYPE_H
 #define CPROVER_SOLVERS_FLATTENING_C_BIT_FIELD_REPLACEMENT_TYPE_H
 
-#include <util/std_types.h>
+#include <util/c_types.h>
 #include <util/namespace.h>
 
 typet c_bit_field_replacement_type(
