+++ src/util/c_types.cpp
@@ -6,12 +6,12 @@ Author: Daniel Kroening, kroening@kroening.com
 
 \*******************************************************************/
 
+#include "c_types.h"
 
-#include "std_types.h"
 #include "config.h"
 #include "invariant.h"
-
-#include "c_types.h"
+#include "pointer_offset_size.h"
+#include "std_types.h"
 
 bitvector_typet index_type()
 {
@@ -302,4 +302,35 @@ std::string c_type_as_string(const irep_idt &c_type)
     return "signed __int128";
   else
     return "";
+}
+
+optionalt<std::pair<struct_union_typet::componentt, mp_integer>>
+union_typet::find_widest_union_component(const namespacet &ns) const
+{
+  const union_typet::componentst &comps = components();
+
+  optionalt<mp_integer> max_width;
+  typet max_comp_type;
+  irep_idt max_comp_name;
+
+  for(const auto &comp : comps)
+  {
+    auto element_width = pointer_offset_bits(comp.type(), ns);
+
+    if(!element_width.has_value())
+      return {};
+
+    if(max_width.has_value() && *element_width <= *max_width)
+      continue;
+
+    max_width = *element_width;
+    max_comp_type = comp.type();
+    max_comp_name = comp.get_name();
+  }
+
+  if(!max_width.has_value())
+    return {};
+  else
+    return std::make_pair(
+      struct_union_typet::componentt{max_comp_name, max_comp_type}, *max_width);
 }
