+++ src/util/format_type.cpp
@@ -7,9 +7,9 @@ Author: Daniel Kroening, kroening@kroening.com
 \*******************************************************************/
 
 #include "format_type.h"
+#include "c_types.h"
 #include "format_expr.h"
 #include "pointer_expr.h"
-#include "std_types.h"
 
 #include <ostream>
