+++ src/util/simplify_expr_int.cpp
@@ -11,6 +11,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "arith_tools.h"
 #include "bitvector_expr.h"
 #include "byte_operators.h"
+#include "c_types.h"
 #include "config.h"
 #include "expr_util.h"
 #include "fixedbv.h"
