+++ src/util/simplify_utils.cpp
@@ -10,6 +10,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "arith_tools.h"
 #include "byte_operators.h"
+#include "c_types.h"
 #include "config.h"
 #include "endianness_map.h"
 #include "namespace.h"
